Tidy up changes and merge rng tests
